[MNT] Dataset downloader testing workflow (#5437)
Replaces #5004
Depends on #5304
