+++ .github/workflows/validate_datasets.yml
@@ -1,0 +1,81 @@
+name: datasets validation
+on:
+  pull_request:
+    branches:
+      - main
+    paths:
+      - sktime/datasets/**
+  schedule:
+    - cron: 0 0 1 * *
+  workflow_dispatch:
+jobs:
+  test_internal_data:
+    name: test data shipped with sktime
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version:
+          - "3.8"
+          - "3.9"
+          - "3.10"
+          - "3.11"
+          - "3.12"
+        operating-system:
+          - macos-latest
+          - ubuntu-latest
+          - windows-latest
+    runs-on: ${{ matrix.operating-system }}
+    steps:
+      - name: repository checkout step
+        uses: actions/checkout@v4
+      - name: update tracking reference step
+        run: git remote set-branches origin main
+        shell: bash
+      - name: shallow clone update step
+        run: git fetch --depth 1
+        shell: bash
+      - name: python environment step
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: dependencies installation step
+        run: python3 -m pip install .[tests]
+        shell: bash
+      - name: unit test step
+        run: python3 -m pytest -m "not datadownload" sktime/datasets
+        shell: bash
+  test_external_data:
+    name: test data downloaded from external sources
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version:
+          - "3.8"
+          - "3.9"
+          - "3.10"
+          - "3.11"
+          - "3.12"
+        operating-system:
+          - macos-latest
+          - ubuntu-latest
+          - windows-latest
+    runs-on: ${{ matrix.operating-system }}
+    steps:
+      - name: repository checkout step
+        uses: actions/checkout@v4
+      - name: update tracking reference step
+        run: git remote set-branches origin main
+        shell: bash
+      - name: shallow clone update step
+        run: git fetch --depth 1
+        shell: bash
+      - name: python environment step
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: dependencies installation step
+        run: python3 -m pip install .[tests]
+        shell: bash
+      - name: unit test step
+        run: python3 -m pytest -m "datadownload" sktime/datasets
+        shell: bash
